changed database from sqlite to postgres, reset migrations, updated docker compose and small changes at the classes

+++ Models/ItemPedido.cs
@@ -5,6 +5,6 @@
         public int ItemPedidoId { get; set; }
         public int ProdutoId { get; set; } // Referência ao produto
         public int Quantidade { get; set; }
-        public decimal ValorTotalItem { get; set; }
+        public Double ValorTotalItem { get; set; }
     }
 }
